+++ source/Examples/PrimaryInstall.ps1
@@ -274,6 +274,20 @@ Configuration PrimaryInstall
             [array]$cmAccountsDependsOn += "[CMAccounts]AddingAccount-$($account.Username)"
         }
 
+        CMEmailNotificationComponent EmailSettings
+        {
+            SiteCode             = $SiteCode
+            SendFrom             = 'emailsender@contoso.com'
+            SmtpServerFqdn       = 'EmailServer.contoso.com'
+            TypeOfAuthentication = 'Other'
+            Port                 = 465
+            UseSsl               = $true
+            Enabled              = $true
+            UserName             = 'contoso\EmailUser'
+            PsDscRunAsCredential = $SccmInstallAccount
+            DependsOn            = $cmAccountsDependsOn
+        }
+
         CMForestDiscovery CreateForestDiscovery
         {
             SiteCode             = $SiteCode
@@ -537,6 +551,33 @@ Configuration PrimaryInstall
             DependsOn                     = '[Script]RebootAfterSccmSetup'
         }
 
+        CMSoftwareUpdatePointComponent SUPComponent
+        {
+            SiteCode                                = $SiteCode
+            EnableSynchronization                   = $true
+            SynchronizeAction                       = 'SynchronizeFromMicrosoftUpdate'
+            ScheduleType                            = 'Days'
+            RecurInterval                           = 7
+            LanguageSummaryDetailsToInclude         = @('English')
+            LanguageUpdateFilesToInclude            = @('English')
+            ProductsToInclude                       = @('Windows 10')
+            UpdateClassificationsToInclude          = @('Critical Updates','Updates')
+            ContentFileOption                       = 'FullFilesOnly'
+            DefaultWsusServer                       = $ServerName
+            EnableCallWsusCleanupWizard             = $true
+            EnableSyncFailureAlert                  = $true
+            ImmediatelyExpireSupersedence           = $false
+            ImmediatelyExpireSupersedenceForFeature = $false
+            ReportingEvent                          = 'DoNotCreateWsusReportingEvents'
+            WaitMonth                               = 1
+            WaitMonthForFeature                     = 1
+            EnableThirdPartyUpdates                 = $true
+            EnableManualCertManagement              = $false
+            FeatureUpdateMaxRuntimeMins             = 300
+            NonFeatureUpdateMaxRuntimeMins          = 300
+            DependsOn                               = '[CMSoftwareUpdatePoint]SUPInstall'
+        }
+
         Script RebootAfterSCCMConfigurationInstall
         {
             TestScript = {
@@ -551,7 +592,7 @@ Configuration PrimaryInstall
                 '[CMHeartbeatDiscovery]CreateHeartbeatDiscovery','[CMUserDiscovery]CreateUserDiscovery','[CMClientStatusSettings]CreateClientStatusSettings',$cmSiteMaintenanceDependsOn,
                 '[CMBoundaryGroups]DemoBoundaryGroup','[CMAdministrativeUser]SiteAdmins','[CMCollectionMembershipEvaluationComponent]CollectionSettings',
                 '[CMStatusReportingComponent]StatusReportingSettings','[Registry]MaxHWMifSize','[CMDistributionPointGroupMembers]DPGroupMembers','[CMManagementPoint]MPInstall',
-                '[CMSoftwareUpdatePoint]SUPInstall'
+                '[CMSoftwareUpdatePoint]SUPInstall','[CMEmailNotificationComponent]EmailSettings','[CMSoftwareUpdatePointComponent]SUPComponent'
         }
     }
 }
@@ -579,6 +620,7 @@ $params = @{
         Get-Credential -Username 'contoso\SCCM-Network' -Message 'SCCM Network Service account'
         Get-Credential -Username 'contoso\SCCM-ClientPush' -Message 'SCCM Client Push account'
         Get-Credential -Username 'contoso\SCCM-ADJoin' -Message 'SCCM AD Join account'
+        Get-Credential -Username 'contoso\EmailUser' -Message 'User account for authenticating against an SMTP server'
     )
 }